make tables that include different fire sequences
the problem is that some runs are missing (for e.g. it could be the fire
series 125 of run 10) so when it reaches this file it does not load
anynmore and it gives an error. The idea would be that it skiped this
file and continue to read the others so that the table is still created.
A bunch of thanks Mara!